Struct rustcoalescence_algorithms_cuda_cpu_kernel::SimulationKernel

pub struct SimulationKernel<M: MathsCore, H: Habitat<M> + RustToCuda, G: PrimeableRng<M> + RustToCuda, S: LineageStore<M, H> + RustToCuda, X: EmigrationExit<M, H, G, S> + RustToCuda, D: DispersalSampler<M, H, G> + RustToCuda, C: CoalescenceSampler<M, H, S> + RustToCuda, T: TurnoverRate<M, H> + RustToCuda, N: SpeciationProbability<M, H> + RustToCuda, E: MinSpeciationTrackingEventSampler<M, H, G, S, X, D, C, T, N> + RustToCuda, I: ImmigrationEntry<M> + RustToCuda, A: SingularActiveLineageSampler<M, H, G, S, X, D, C, T, N, E, I> + RustToCuda, ReportSpeciation: Boolean, ReportDispersal: Boolean> { /* private fields */ }

Implementations

impl<M: MathsCore, H: Habitat<M> + RustToCuda, G: PrimeableRng<M> + RustToCuda, S: LineageStore<M, H> + RustToCuda, X: EmigrationExit<M, H, G, S> + RustToCuda, D: DispersalSampler<M, H, G> + RustToCuda, C: CoalescenceSampler<M, H, S> + RustToCuda, T: TurnoverRate<M, H> + RustToCuda, N: SpeciationProbability<M, H> + RustToCuda, E: MinSpeciationTrackingEventSampler<M, H, G, S, X, D, C, T, N> + RustToCuda, I: ImmigrationEntry<M> + RustToCuda, A: SingularActiveLineageSampler<M, H, G, S, X, D, C, T, N, E, I> + RustToCuda, ReportSpeciation: Boolean, ReportDispersal: Boolean> SimulationKernel<M, H, G, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>

pub fn try_new( stream: Stream, grid: GridSize, block: BlockSize, ptx_jit: bool, on_compile: Box<KernelCompilationCallback> ) -> CudaResult<Self>where Self: SimulatableKernel<M, H, G, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>,

Errors

Returns a CudaError if loading the CUDA kernel failed.